requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
PyMySQL
Flask-JWT-Extended
Celery
redis
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
cryptography
flask-cors

